+++ tests/test_lists.py
@@ -12,8 +12,28 @@ def test_repr():
     assert model.objects.__repr__() == "ObjList [2 objects]"
     l1 = model.agents.id
     l2 = l1 + 1
-    assert l1.__repr__() == "AttrList of attribute 'id': [1]"
+    assert l1.__repr__() == "AttrList of 'id': [1]"
     assert l2.__repr__() == "AttrList: [2]"
+
+
+def test_call():
+    class MyAgent(ap.Agent):
+        def method(self):
+            if self.id == 2:
+                self.model.agents[2].delete()
+            self.model.called.append(self.id)
+
+    model = ap.Model()
+    model.called = []
+    model.add_agents(4, MyAgent)
+    model.agents.call('method', check_alive=True)
+    assert model.called == [1, 2, 4]
+
+    model = ap.Model()
+    model.called = []
+    model.add_agents(4, MyAgent)
+    model.agents.method()
+    assert model.called == [1, 2, 3, 4]
 
 
 def test_attr_calls():
@@ -24,7 +44,7 @@ def test_attr_calls():
     assert list(model.agents.x) == [1, 1]
     assert list(model.agents.f()) == [2, 2]
     with pytest.raises(AttributeError):
-        assert model.agents.y
+        assert list(model.agents.y)  # Convert to list to call attribute
     with pytest.raises(TypeError):
         assert model.agents.x()  # noqa
 
@@ -45,8 +65,7 @@ def test_select():
     assert selection4 == [False, False, True]
     assert selection5 == [True, True, False]
     assert selection6 == [False, True, True]
-    assert model.agents(selection1) == model.agents.select(selection1)
-    assert list(model.agents(selection1).id) == [2]
+    assert list(model.agents.select(selection1).id) == [2]
 
 
 def test_random():