+++ docs/changelog.rst
@@ -4,6 +4,56 @@
 Changelog
 =========
 
+Please note that API changes without backward compatibility
+are still possible until the release of v1.0.0.
+
+0.0.8.dev0
+----------
+
+Improved lists
+..............
+
+Agent lists have a new method :func:`AgentList.call` that can be used to call
+a method for each agent in the list and has two optional arguments:
+`check_alive`, which prevents the call of agents that are deleted during the call loop;
+and `iter_kwargs`, which can be used to pass method arguments that are different for
+each agent.
+
+The structure of :class:`AttrList` has been changed to an iterable over its
+source list. This improves performance and makes it possible to change
+agent attributes by setting new values to items in the attribute list (see
+:class:`AgentList` for an example). Otherwise, the class behaves as before.
+
+Agent-Environment interaction
+.............................
+
+The method :func:`Agent.neighbors` can now take multiple environments
+as an argument. If no environments are given, it will select all of the
+agents environments. For environments without a topology like :class:`Environment`
+and agents without any environments, an empty list will be returned.
+Agents further have a new attribute :obj:`alive` that indicates whether they
+have been removed from the model.
+
+Other changes
+.............
+
+Following Python's philosophy that explicit is better than implicit,
+a few 'magic' features have been removed:
+
+* Calling :func:`AgentList.select` through `AgentList.__call__`
+  (i.e. by invoking the agent list as a function) is no longer possible.
+* :obj:`Agent.env` and :obj:`Model.env` now returns an error if the agent has no or
+  more than one environment.
+* The argument `env` in :func:`AgentList.position`, :func:`AgentList.move_to`,
+  :func:`AgentList.move_by`, and :func:`AgentList.exit` must be given explicitely
+  if the agent has more than one environment.
+
+Fixes
+.....
+
+* :func:`Model.remove_agents` now removes agents not just from the model,
+  but also from all environments.
+
 0.0.7 (March 2021)
 ------------------
 
@@ -31,8 +81,8 @@ and has three new arguments 'replace', 'weights', and 'shuffle'.
 More information with examples can be found in the API reference
 and the new user guide :doc:`guide_random`.
 
-API changes
-...........
+Other changes
+.............
 
 * The function :func:`sensitivity_sobol` now has an argument :attr:`calc_second_order` (default False).
   If True, the function will add second-order indices to the output.